chore: move metrics test to dedicated _test package
